revised some parts so it makes sense
